+++ src/Data/Contact.php
@@ -2,7 +2,6 @@
 
 namespace Samwilson\Twyne\Data;
 
-use Exception;
 use Samwilson\Twyne\Database;
 
 class Contact extends ModelBase
@@ -18,22 +17,59 @@ class Contact extends ModelBase
         return 'c';
     }
 
-    public static function getByUserAndName(User $user, string $name)
+    public static function newForUser(User $user): Contact
     {
-        $sql = 'SELECT * FROM contacts WHERE c_user = :user AND c_name = :name';
-        $contactData = Database::getInstance()->query($sql, ['user' => $user->getId(), 'name' => $name])->fetch();
-        if (!$contactData) {
-            throw new Exception('Unable to find contact "' . $name . '"');
+        return new self((object)[
+            'c_user' => $user->getId(),
+        ]);
+    }
+
+    public function canBeEditedBy(?User $user): bool
+    {
+        if (!$user) {
+            return false;
         }
-        return new static($contactData);
+        $userExists = (bool)$user->getId();
+        $contactExists = (bool)$this->getId();
+        $contactBelongsToUser = $this->getUserId() === $user->getId();
+        return $userExists && ($contactExists && $contactBelongsToUser || !$contactExists);
+    }
+
+    public function canBeViewedBy(?User $user)
+    {
+        $hasPosts = count($this->getPosts()) > 0;
+        return $hasPosts || $this->canBeEditedBy($user);
     }
 
     /**
-     * @return string|null
+     * @return Post[]
      */
-    public function getName()
+    public function getPosts()
     {
-        return $this->data->c_name;
+        $postData = $this->db->query('SELECT * FROM posts WHERE p_author = :id', ['id' => $this->getId()]);
+        $posts = [];
+        foreach ($postData as $post) {
+            $posts[] = new Post((object)$post);
+        }
+        return $posts;
+    }
+
+    public static function getByUserAndName(User $user, string $name)
+    {
+        $sql = 'SELECT * FROM contacts WHERE c_user = :user AND c_name LIKE :name';
+        $contactData = Database::getInstance()->query($sql, ['user' => $user->getId(), 'name' => $name])->fetch();
+        if (!$contactData) {
+            $contactData = [
+                'c_user' => $user->getId(),
+                'c_name' => $name,
+            ];
+        }
+        return new static((object)$contactData);
+    }
+
+    public function getName(): ?string
+    {
+        return $this->data->c_name ?? null;
     }
 
     /**
@@ -42,6 +78,29 @@ class Contact extends ModelBase
     public function setName(string $name): void
     {
         $this->data->c_name = trim($name);
+    }
+
+    public function getDescription(): ?string
+    {
+        return $this->data->c_description ?? null;
+    }
+
+    /**
+     * @param string $description
+     */
+    public function setDescription(string $description): void
+    {
+        $this->data->c_description = trim($description);
+    }
+
+    public function getUserId(): ?int
+    {
+        return isset($this->data->c_user) ? (int)$this->data->c_user : null;
+    }
+
+    public function getUser(): ?User
+    {
+        return User::loadById($this->getUserId());
     }
 
     public function setUser(User $user)