+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Users can register, get password reminders, and log in and out.
 * Users can create and edit posts.
+* Users can maintain a list of their contacts.
+* Posts can use any of a user's contacts as an author.
 
 [Unreleased]: https://github.com/samwilson/twyne/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.1.0
+[0.1.0]: https://github.com/samwilson/twyne/releases/tag/v0.1.0